Move the filters actions to a custom hook.
